pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pywiggum"
version = "0.1.0"
description = "AI Agent Orchestrator with Dashboard - oversee your autonomous coding agents"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "PyWiggum Contributors" }
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-pyyaml",
]

[project.scripts]
pywiggum = "pywiggum.cli:main"

[project.urls]
Homepage = "https://github.com/pywiggum/pywiggum"
Repository = "https://github.com/pywiggum/pywiggum"
Issues = "https://github.com/pywiggum/pywiggum/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pywiggum"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
